refactor: PyPI CloudVolume, Storage, ThreadedQueue, and lib
Start using the above features from the PyPI library cloud-volume.
Also clean up a few old scripts that have outlived their usefulness.
