--- python/neuroglancer/pipeline/volumes/__init__.py
@@ -1,6 +1,0 @@
-from volumes import *
-from cloudvolume import *
-
-
-
-
