+++ python/requirements.txt
@@ -1,19 +1,11 @@
 backports.lzma==0.0.8
-boto==2.39.0
 blosc==1.5.0
 click==6.6
 Cython==0.25.2
-google-api-python-client==1.6.2
-google-cloud==0.27.0
-intern==0.9.4
-jsonschema==2.6.0
+cloud-volume==0.2.0
 networkx==1.11
 numpy==1.13.1
-Pillow==4.2.1
-protobuf==3.3.0
 pytest==3.0.7 
-python-jsonschema-objects==0.2.1
-requests==2.18.4
 sockjs-tornado==1.0.3
 tenacity==4.4.0
 tornado==4.4.2
