--- python/test/test_lib.py
@@ -1,34 +1,0 @@
-import pytest
-
-import neuroglancer.lib as lib
-
-def test_divisors():
-
-  divisors = (
-    (1, (1,)), 
-    (2, (1,2)),
-    (3, (1,3)),
-    (4, (1,2,4)),
-    (6, (1,2,3,6)),
-    (35, (1,5,7,35)),
-    (128, (1,2,4,8,16,32,64,128)),
-    (258, (1,2,3,6,43,86,129,258)),
-  )
-
-  for num, ans in divisors:
-    result = [ _ for _ in lib.divisors(num) ]
-    result.sort()
-    assert tuple(result) == ans
-
-def test_find_closest_divisor():
-  size = lib.find_closest_divisor( (128,128,128), (64,64,64) )
-  assert tuple(size) == (64,64,64)
-
-  size = lib.find_closest_divisor( (240,240,240), (64,64,64) )
-  assert tuple(size) == (60,60,60)
-
-  size = lib.find_closest_divisor( (224,224,224), (64,64,64) )
-  assert tuple(size) == (56,56,56)
-
-  size = lib.find_closest_divisor( (73,73,73), (64,64,64) )
-  assert tuple(size) == (73,73,73)
